# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API – Teste Devs UEFS",
        "description": "API RESTful para gerenciamento de usuários, posts e tags.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Ambiente local"
        }
    ],
    "paths": {
...
        "/api/v1/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Lista tags",
                "operationId": "e320cb59dba3e323dacbbc502509717c",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filtro por nome ou slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de tags"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Cria uma nova tag",
                "operationId": "d5b48558ae6b21f20e758584734e2cd7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tecnologia da Informação"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "tecnologia-da-informacao"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tags/{id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Detalha uma tag",
                "operationId": "a3d7580f0301bd6fff879d635b82a2fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag encontrada"
                    },
                    "404": {
                        "description": "Tag não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Edita uma tag",
                "operationId": "ab9fc572397025dcf3a29db57acdebdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tecnologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tag atualizada"
                    },
                    "404": {
                        "description": "Tag não atualizada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Deleta uma tag",
                "operationId": "fab7e3dec2348caad78c23998aa54ba3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tag deletada"
                    },
                    "404": {
                        "description": "Tag não deletada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token JWT ou Sanctum no formato: Bearer {token}"
            }
        }
    },
...
}